add copyright string
