+++ client/JSDebugger/mlRuntime.ts
@@ -1,8 +1,50 @@
+/*
+ * Copyright (c) 2020 MarkLogic Corporation
+ */
+
 import { EventEmitter } from 'events';
 //@ts-ignore
 import * as request from 'request-promise';
 import * as fs from 'fs';
 import * as querystring from 'querystring';
+
+/**Interfaces defined for Debugger Response*/
+
+export interface V8Frame {
+	callFrameId: string;
+	functionName?: string;
+	functionLocation?: object;
+	location: {
+		scriptId:string,
+		lineNumber:number,
+		columnNumber:number
+	};
+	url: string;
+	scopeChain: ScopeObject[];
+	this: V8PropertyValue;
+}
+
+export interface ScopeObject {
+	type: string;
+	object: V8PropertyValue;
+}
+
+export interface V8PropertyObject {
+	name: string;
+	value: V8PropertyValue;
+	writable?: boolean;
+	configurable?: boolean;
+	enumerable?:boolean;
+	isOwn?:boolean;
+}
+
+export interface V8PropertyValue {
+	type: string;
+	value?: any;
+	classname?: string;
+	description?: string;
+	objectId?: string;
+}
 
 export interface MLbreakPoint {
 	url:string;
